requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
apscheduler==3.10.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0